refactor: Shuffle around a bit
