+++ common/src/main/java/com/wynntils/services/mapdata/attributes/resolving/DerivedMapVisibility.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.services.mapdata.attributes;
+package com.wynntils.services.mapdata.attributes.resolving;
 
 import com.wynntils.services.mapdata.attributes.type.MapVisibility;
 import java.util.Optional;
